GDAL provider: maskband option: generate alpha channel from dataset mask
Add a 'maskband' option to the GDAL dataset provider which, if
specified, gives the 1-based raster band index whose mask should be used
as an alpha channel.

This allows datasets with 'NoData' values to be handled nicely on the
map. It also has the nice side-effect of causing pixels outside of the
raster to become transparent.

Unfortunately this does reduce the efficiency of the provider in that
the raster has to be re-projected twice.