Add 'Access-Control-Allow-Origin' to header in requestHandler2
In requestHandler2, the allowed_origin parameter for a given layer is successfully parsed, and inserted into an instance of wsgiref.headers Headers class.

However the variable `headers` is overwritten upon a successful return from `layer.getTileResponse()`. This removes the value a user has provided for the "allowed origins" parameter for a layer within the TileStache config.