Improved performance of _addRecentTile
Deleting items from the list without an index will
have to search the whole list, so the loop will have
O(n^2) performance. Instead, we just remember the index
of the first tile that is still new enough and then
cut off the list at that point. Performance should now be O(n).

Note that it might be even better (more transparent)
to use an OrderedDict here.